requirements.txt
cohere==1.3.2
feedparser==6.0.10
beautifulsoup4==4.9.1
python-dateutil==2.8.2
requests==2.31.0
article-parser==1.6.0
flask==3.0.0
Markdown2==2.4.10
parse==1.19.0